+++ src/section/campaign/campaignUpdateValidateAndPrepare.ts
@@ -48,7 +48,7 @@ const campaignUpdateValidateAndPrepare = async (intl: IntlShape<string>, isDev: 
     updateSet = { ...updateSet, budget: data.budget };
   }
   //
-  updateCall.parameter = `$id_${updateCall.idx}: bigint!, $p_${updateCall.idx}: campaign_set_input`;
+  updateCall.parameter = `$id_${updateCall.idx}: Int!, $p_${updateCall.idx}: campaign_set_input`;
   updateCall.command = `
     data: update_campaign_by_pk(pk_columns: {id: $id_${updateCall.idx}}, _set: $p_${updateCall.idx}) {
       id
