+++ src/section/campaign/query.ts
@@ -2,7 +2,7 @@ import { executeGraphql } from "../../db/util";
 
 export const getCampaignTypeById = async (id: number) => {
   return await executeGraphql(`
-    query ($id: bigint!) {
+    query ($id: Int!) {
       data:campaign_type_by_pk(id: $id) {
         id
       }
@@ -12,7 +12,7 @@ export const getCampaignTypeById = async (id: number) => {
 }
 
 export const getCampaignById = async (id: number) => {
-  return await executeGraphql(`query ($id: bigint!) {
+  return await executeGraphql(`query ($id: Int!) {
     data:campaign_by_pk(id: $id){
       id
       organization_id
